# nvgpu optimize reshape v3