--- a/src/ngraph/runtime/gpu/cuda_emitter.hpp
+++ b/src/ngraph/runtime/gpu/cuda_emitter.hpp
@@ -154,6 +154,10 @@ namespace ngraph
                                      NVShape input_shape,
                                      NVShape input_order);
 
+                size_t build_reshape_2d(const std::array<std::string, 2>& dtypes,
+                                        NVShape input_shape,
+                                        NVShape input_order);
+
                 size_t build_reshape_3d(const std::array<std::string, 2>& dtypes,
                                         NVShape input_shape,
                                         NVShape input_order);