… src/ngraph/runtime/gpu/gpu_emitter.cpp
… src/ngraph/runtime/intelgpu/intelgpu_op_batchnorm.cpp
… test/backend_test.in.cpp