Another 3.7x compare to previous optimization. 
2,968us -> 779us

changes:
1.  remove dimension with size 1.
2. combine dimensions that's still in order after reorder
3. add 2D/3D shared memory tiled reshape. 
4. 2D reshape is 1.1x - 1.4x faster than cublas.
5. 3D reshape is 4 - 6x faster than previous optimization


many reshape can be use 2D/3D shared memory tiled reshape with cuda now, both are 4-6x faster.
for example: {1, 1, 1024, 256} -> {256, 1024, 1, 1} with input order {3, 2, 0, 1}
      become: {1024, 256} -> {256, 1024} with input order {1, 0}
which can use cublas matrix transpose.

for example: {256, 256, 3, 3} -> {3, 3, 256, 256} with input order {3, 2, 0, 1}
      become: {256, 256, 9} -> {9, 256, 256} with input order {2, 1, 0}
which can use 3D tiled reshape.

TODO: nd shared memory tiled reshape

```
Benchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,043ms
11.5ms per iteration

---- Aggregate times per op type ----
Convolution     7,766us
Reshape         2,968us
BatchNorm         497us
Relu              334us
Add               202us
MaxPool            59us
Dot                34us
Sum                13us
Pad                10us
Divide              5us
Broadcast           5us
Result              1us
```

after:
```
enchmarking ../ngraph-models/models/tensorflow/inference/resnet50/tf_function_ngraph_cluster_rn50_bs1.json, GPU backend, 10 iterations.
compile time: 7,057ms
8ms per iteration

---- Aggregate times per op type ----
Convolution     6,565us
Reshape           779us
BatchNorm         463us
Relu              291us
Add               191us
MaxPool            53us
Dot                33us
Sum                11us
Pad                 8us
Divide              5us
Broadcast           4us
Result              2us
```