@fengleitian 2018-09-17
Added tests to cover 3D/2D reshape.